trino/src/requirements.txt
polars==0.20.31
duckdb==1.1.0
psycopg2-binary==2.9.9
clickhouse-connect==0.7.19
minio==7.2.7
pyarrow==17.0.0

